+++ CHANGELOG.md
@@ -26,6 +26,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - fetch_slv:  Fixed incorrect reference to slave name variable.
 
 ### Changed
+- main:  Added ProgramLock class to implement program locking.
 - config/slave.txt.TEMPLATE:  Changed format of file.
 - main:  Refactored the "if" statements to streamline the checks.
 - run_program:  Changed variables to standard naming convention.